global exception added

+++ src/main/java/com/usermgm/service/userService.java
@@ -1,7 +1,6 @@
 package com.usermgm.service;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -14,11 +13,13 @@ import com.usermgm.repository.userRepository;
 @Service
 public class userService {
 
+	
+	//dependency injection for repository
 	@Autowired
 	private userRepository userepo;	
 
 	
-	
+	//adding the new user which satisfies user validation
 	public user addUser(userDTO u1) {
 	
 		user u2=user.build(u1.getUserId(), u1.getUserName(), u1.getUserPhone(), u1.getUserEmail(),u1.getUserPassword());
@@ -28,26 +29,46 @@ public class userService {
 		
 	}
 	
+	//retriving all the users 
 	public List<user> getUser(){
 		return userepo.findAll();
 	}
 	
 	
-	public String deleteUser(int id)  {
+	//deleting the already existed user
+	//if user is not existed we are throwing globalException
+	public user deleteUser(int id) throws globalException  {
 	
-		userepo.deleteById(id);
-		return "Deleted!...";
+		user _u2=userepo.findByUserId(id);
+		if(_u2!=null) {
+			userepo.deleteById(id);
+			return _u2;
+		}
+		else {
+			throw new globalException("user not found!...");
+		}
+		
 	}
 	
+	//updating user details
 	public user updateUser(userDTO u1) {
 		
-user u2=user.build(u1.getUserId(), u1.getUserName(), u1.getUserPhone(), u1.getUserEmail(),u1.getUserPassword());
-		
-		
+		user u2=user.build(u1.getUserId(), u1.getUserName(), u1.getUserPhone(), u1.getUserEmail(),u1.getUserPassword());
 		return userepo.save(u2);
 	}
 	
 	
+	//retriving details of specific user
+	public user getUser(int userid) throws globalException
+	{
+		user _u1=userepo.findByUserId(userid);
+		if(_u1!=null) {
+			return _u1;
+		}
+		else {
+			throw new globalException("user not found with userid "+userid);
+		}
+	}
